refactor: group rest of components
