+++ frontend/src/components/MainView/index.tsx
@@ -4,8 +4,8 @@ import { useMapTransform } from "../../hooks/useMapTransform";
 import { useRoomProperties } from "../../hooks/useRoomProperties";
 import { useRoomSelection } from "../../hooks/useRoomSelection";
 import Exactum2 from "../assets/exactum-2.min.svg?react";
-import ColorToggle from "../ColorToggle";
 import SidePanel from "../Sidepanel";
+import ColorToggle from "./ColorToggle";
 import Legend from "./Legend";
 import "./MainView.css";
 import ZoomButtons from "./ZoomButtons";
