+++ frontend/src/components/PersonModal/PersonForm/index.tsx
@@ -7,7 +7,7 @@ import {
   type ReferenceItem,
 } from "../../../services/referenceDataService";
 import type { Person } from "../../../types";
-import PersonSelector from "../../PersonSelector";
+import PersonSelector from "../../PersonSelector/PersonSelector";
 import "./PersonForm.css";
 
 interface FieldDef {
