+++ frontend/src/components/PersonSearch/style.css
