+++ frontend/src/components/PersonSelector/index.tsx
@@ -1,4 +1,4 @@
-import type { Person } from "../types";
+import type { Person } from "../../types";
 import "./PersonSelector.css";
 
 interface PersonSelectorProps {
