+++ frontend/src/components/PersonSelector/style.css
