+++ frontend/src/components/RoomSelectionProvider/index.tsx
@@ -2,9 +2,9 @@ import { useEffect, useRef, useState } from "react";
 import type {
   ExpandReq,
   RoomSelectionContextType,
-} from "../contexts/RoomSelectionContext";
-import { RoomSelectionContext } from "../contexts/RoomSelectionContext";
-import type { Room } from "../types";
+} from "../../contexts/RoomSelectionContext";
+import { RoomSelectionContext } from "../../contexts/RoomSelectionContext";
+import type { Room } from "../../types";
 
 const SKELETON_DELAY_MS = 250;
