+++ frontend/tests/components/mainView.test.tsx
@@ -4,7 +4,7 @@ import userEvent from "@testing-library/user-event";
 import type { ReactElement } from "react";
 import { describe, expect, it, vi } from "vitest";
 import MainView from "../../src/components/MainView/index.tsx";
-import { RoomSelectionProvider } from "../../src/components/RoomSelectionProvider.tsx";
+import { RoomSelectionProvider } from "../../src/components/RoomSelectionProvider/index.tsx";
 import {
   DEFAULT_SCALE,
   getBottomBound,
